refactor of received namespace, moved to folder receipts
